サイドバーとグラフを作りました
